Moved doPagePaste into a test helper file, ready for re-use.
